+++ config/config.js
@@ -9,7 +9,15 @@ const config = {
   dbName: process.env.DB_NAME,
   dbPort: process.env.DB_PORT,
   jwtLogin: process.env.JWT_LOGIN,
-  jwtPassword: process.env.JWT_MAIL
+  jwtPassword: process.env.JWT_MAIL,
+  firebase: {
+    apiKey:process.env.FR_APIKEY,
+    authDomain:process.env.FR_AUTHDOMAIN,
+    projectId:process.env.FR_PROJECTID,
+    storageBucket:process.env.FR_BUCKET,
+    messagingSenderId:process.env.FR_SENDERID,
+    appId:process.env.FR_APP_ID
+  }
 }
 
 module.exports = { config };