+++ middleware/multer/multer.middleware.js
@@ -1,0 +1,9 @@
+const Multer = require('multer');
+
+const multer = Multer({
+    storage: Multer.memoryStorage()
+});
+
+module.exports = {
+    multer
+}